refactor: Extract copy to clipboard function
